[Javadoc] Wrote class and method level comments
Wrote class and method level comments for Element, Empty, Immovable and its subclasses, and some of Movable. Changed density to use an integer instead of a float to conserve a small bit of memory.

+++ PixelPhysics/core/src/com/aidenx11/game/elements/movable/gas/Gas.java
@@ -7,7 +7,7 @@ public class Gas extends Movable {
 
 	public Gas(ElementTypes type, int row, int column, CustomColor color, boolean canDie, int lifetime,
 			boolean flammable, boolean extinguishesThings, float chanceToCatch, boolean burnsThings, float velocity,
-			float acceleration, float maxSpeed, float density, boolean movesSideways) {
+			float acceleration, float maxSpeed, int density, boolean movesSideways) {
 		super(type, row, column, color, canDie, lifetime, flammable, extinguishesThings, chanceToCatch, burnsThings,
 				velocity, acceleration, maxSpeed, density, movesSideways, false, 0f, 0);
 		super.setFreeFalling(true);
