+++ PixelPhysics/core/src/com/aidenx11/game/elements/movable/liquid/Water.java
@@ -17,7 +17,7 @@ public class Water extends Liquid {
 	public static ElementTypes type = ElementTypes.WATER;
 	private static float acceleration = pixelPhysicsGame.GRAVITY_ACCELERATION + 0.2f;
 	private static float maxSpeed = 15f;
-	private static float density = 6f;
+	private static int density = 6;
 	private static int dispersionRate = 8;
 
 	int[] darkerWater = new int[] { 15, 94, 156 };
